# internal/server/swagger/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /v1/member/invite/email:
        post:
            tags:
                - Member
            operationId: Member_InviteMemberFromEmail
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.InviteMemberFromEmailRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.InviteMemberReply'
    /v1/member/invite/phone:
        post:
            tags:
                - Member
            operationId: Member_InviteMemberFromPhone
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.InviteMemberFromPhoneRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.InviteMemberReply'
    /v1/member/terminate/{uid}:
        delete:
            tags:
                - Member
            operationId: Member_TerminateMember
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.TerminateMemberReply'
    /v1/member/{uid}:
        get:
            tags:
                - Member
            operationId: Member_GetMember
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.MemberItem'
    /v1/member/{uid}/status:
        put:
            tags:
                - Member
            operationId: Member_UpdateMemberStatus
            parameters:
                - name: uid
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/goddess.api.v1.UpdateMemberStatusRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.UpdateMemberStatusReply'
    /v1/members:
        get:
            tags:
                - Member
            operationId: Member_ListMember
            parameters:
                - name: page
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: pageSize
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.ListMemberReply'
    /v1/members/select:
        get:
            tags:
                - Member
            operationId: Member_SelectMember
            parameters:
                - name: keyword
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
                - name: lastUID
                  in: query
                  schema:
                    type: string
                - name: status
                  in: query
                  schema:
                    type: integer
                    format: enum
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/goddess.api.v1.SelectMemberReply'
...
components:
    schemas:
...
        goddess.api.v1.InviteMemberFromEmailRequest:
            type: object
            properties:
                email:
                    type: string
        goddess.api.v1.InviteMemberFromPhoneRequest:
            type: object
            properties:
                phone:
                    type: string
        goddess.api.v1.InviteMemberReply:
            type: object
            properties: {}
        goddess.api.v1.ListMemberReply:
            type: object
            properties:
                total:
                    type: string
                page:
                    type: integer
                    format: int32
                pageSize:
                    type: integer
                    format: int32
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.MemberItem'
...
        goddess.api.v1.MemberItem:
            type: object
            properties:
                uid:
                    type: string
                namespaceUID:
                    type: string
                userUID:
                    type: string
                name:
                    type: string
                email:
                    type: string
                phone:
                    type: string
                avatar:
                    type: string
                createdAt:
                    type: string
                updatedAt:
                    type: string
                status:
                    type: integer
                    format: enum
...
        goddess.api.v1.SelectMemberReply:
            type: object
            properties:
                items:
                    type: array
                    items:
                        $ref: '#/components/schemas/goddess.api.v1.MemberItem'
                total:
                    type: string
                lastUID:
                    type: string
                hasMore:
                    type: boolean
...
        goddess.api.v1.TerminateMemberReply:
            type: object
            properties: {}
        goddess.api.v1.UpdateMemberStatusReply:
            type: object
            properties: {}
        goddess.api.v1.UpdateMemberStatusRequest:
            type: object
            properties:
                uid:
                    type: string
                status:
                    type: integer
                    format: enum
...