# pkg/api/v1/member_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.0
// - protoc             v5.29.3
// source: api/v1/member.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMemberGetMember = "/goddess.api.v1.Member/GetMember"
const OperationMemberInviteMemberFromEmail = "/goddess.api.v1.Member/InviteMemberFromEmail"
const OperationMemberInviteMemberFromPhone = "/goddess.api.v1.Member/InviteMemberFromPhone"
const OperationMemberListMember = "/goddess.api.v1.Member/ListMember"
const OperationMemberSelectMember = "/goddess.api.v1.Member/SelectMember"
const OperationMemberTerminateMember = "/goddess.api.v1.Member/TerminateMember"
const OperationMemberUpdateMemberStatus = "/goddess.api.v1.Member/UpdateMemberStatus"

type MemberHTTPServer interface {
	GetMember(context.Context, *GetMemberRequest) (*MemberItem, error)
	InviteMemberFromEmail(context.Context, *InviteMemberFromEmailRequest) (*InviteMemberReply, error)
	InviteMemberFromPhone(context.Context, *InviteMemberFromPhoneRequest) (*InviteMemberReply, error)
	ListMember(context.Context, *ListMemberRequest) (*ListMemberReply, error)
	SelectMember(context.Context, *SelectMemberRequest) (*SelectMemberReply, error)
	TerminateMember(context.Context, *TerminateMemberRequest) (*TerminateMemberReply, error)
	UpdateMemberStatus(context.Context, *UpdateMemberStatusRequest) (*UpdateMemberStatusReply, error)
}

func RegisterMemberHTTPServer(s *http.Server, srv MemberHTTPServer) {
	r := s.Route("/")
	r.POST("/v1/member/invite/email", _Member_InviteMemberFromEmail0_HTTP_Handler(srv))
	r.POST("/v1/member/invite/phone", _Member_InviteMemberFromPhone0_HTTP_Handler(srv))
	r.DELETE("/v1/member/terminate/{uid}", _Member_TerminateMember0_HTTP_Handler(srv))
	r.PUT("/v1/member/{uid}/status", _Member_UpdateMemberStatus0_HTTP_Handler(srv))
	r.GET("/v1/member/{uid}", _Member_GetMember0_HTTP_Handler(srv))
	r.GET("/v1/members", _Member_ListMember0_HTTP_Handler(srv))
	r.GET("/v1/members/select", _Member_SelectMember0_HTTP_Handler(srv))
}

func _Member_InviteMemberFromEmail0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in InviteMemberFromEmailRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberInviteMemberFromEmail)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.InviteMemberFromEmail(ctx, req.(*InviteMemberFromEmailRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*InviteMemberReply)
		return ctx.Result(200, reply)
	}
}

func _Member_InviteMemberFromPhone0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in InviteMemberFromPhoneRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberInviteMemberFromPhone)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.InviteMemberFromPhone(ctx, req.(*InviteMemberFromPhoneRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*InviteMemberReply)
		return ctx.Result(200, reply)
	}
}

func _Member_TerminateMember0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in TerminateMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberTerminateMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.TerminateMember(ctx, req.(*TerminateMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TerminateMemberReply)
		return ctx.Result(200, reply)
	}
}

func _Member_UpdateMemberStatus0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdateMemberStatusRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberUpdateMemberStatus)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateMemberStatus(ctx, req.(*UpdateMemberStatusRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UpdateMemberStatusReply)
		return ctx.Result(200, reply)
	}
}

func _Member_GetMember0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		if err := ctx.BindVars(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberGetMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMember(ctx, req.(*GetMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MemberItem)
		return ctx.Result(200, reply)
	}
}

func _Member_ListMember0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in ListMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberListMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.ListMember(ctx, req.(*ListMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*ListMemberReply)
		return ctx.Result(200, reply)
	}
}

func _Member_SelectMember0_HTTP_Handler(srv MemberHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SelectMemberRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMemberSelectMember)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SelectMember(ctx, req.(*SelectMemberRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SelectMemberReply)
		return ctx.Result(200, reply)
	}
}

type MemberHTTPClient interface {
	GetMember(ctx context.Context, req *GetMemberRequest, opts ...http.CallOption) (rsp *MemberItem, err error)
	InviteMemberFromEmail(ctx context.Context, req *InviteMemberFromEmailRequest, opts ...http.CallOption) (rsp *InviteMemberReply, err error)
	InviteMemberFromPhone(ctx context.Context, req *InviteMemberFromPhoneRequest, opts ...http.CallOption) (rsp *InviteMemberReply, err error)
	ListMember(ctx context.Context, req *ListMemberRequest, opts ...http.CallOption) (rsp *ListMemberReply, err error)
	SelectMember(ctx context.Context, req *SelectMemberRequest, opts ...http.CallOption) (rsp *SelectMemberReply, err error)
	TerminateMember(ctx context.Context, req *TerminateMemberRequest, opts ...http.CallOption) (rsp *TerminateMemberReply, err error)
	UpdateMemberStatus(ctx context.Context, req *UpdateMemberStatusRequest, opts ...http.CallOption) (rsp *UpdateMemberStatusReply, err error)
}

type MemberHTTPClientImpl struct {
	cc *http.Client
}

func NewMemberHTTPClient(client *http.Client) MemberHTTPClient {
	return &MemberHTTPClientImpl{client}
}

func (c *MemberHTTPClientImpl) GetMember(ctx context.Context, in *GetMemberRequest, opts ...http.CallOption) (*MemberItem, error) {
	var out MemberItem
	pattern := "/v1/member/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberGetMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberHTTPClientImpl) InviteMemberFromEmail(ctx context.Context, in *InviteMemberFromEmailRequest, opts ...http.CallOption) (*InviteMemberReply, error) {
	var out InviteMemberReply
	pattern := "/v1/member/invite/email"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMemberInviteMemberFromEmail))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberHTTPClientImpl) InviteMemberFromPhone(ctx context.Context, in *InviteMemberFromPhoneRequest, opts ...http.CallOption) (*InviteMemberReply, error) {
	var out InviteMemberReply
	pattern := "/v1/member/invite/phone"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMemberInviteMemberFromPhone))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberHTTPClientImpl) ListMember(ctx context.Context, in *ListMemberRequest, opts ...http.CallOption) (*ListMemberReply, error) {
	var out ListMemberReply
	pattern := "/v1/members"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberListMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberHTTPClientImpl) SelectMember(ctx context.Context, in *SelectMemberRequest, opts ...http.CallOption) (*SelectMemberReply, error) {
	var out SelectMemberReply
	pattern := "/v1/members/select"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberSelectMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberHTTPClientImpl) TerminateMember(ctx context.Context, in *TerminateMemberRequest, opts ...http.CallOption) (*TerminateMemberReply, error) {
	var out TerminateMemberReply
	pattern := "/v1/member/terminate/{uid}"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMemberTerminateMember))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "DELETE", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

func (c *MemberHTTPClientImpl) UpdateMemberStatus(ctx context.Context, in *UpdateMemberStatusRequest, opts ...http.CallOption) (*UpdateMemberStatusReply, error) {
	var out UpdateMemberStatusReply
	pattern := "/v1/member/{uid}/status"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMemberUpdateMemberStatus))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}